move upload_package from nanny to worker
